refactor : PreLoad를 위한 Builder 위치 변경 #PAR-126
buildExoPlayer는 Battle에 관한 메인 Screen Composable이 만들어질 때 호출이 되고, 다이얼로그를 띄울 때 buildPlayerView 호출함으로써 별도의 ExoPlayer의 빌드 과정이 없이 Player를 보여주도록 한다.
