+++ feature/battle/build.gradle.kts
@@ -15,5 +15,7 @@ dependencies {
     implementation(libs.okhttp.sse)
     implementation("com.google.code.gson:gson:2.10.1")
 
+    implementation(libs.exo.player)
+
     implementation(project(":feature:match"))
 }
